fix: testing strategy moved to github actions
