+++ .pre-commit-config.yaml
@@ -86,12 +86,3 @@ repos:
     rev: 'v0.971'
     hooks:
       - id: mypy
-
-  - repo: local
-    hooks:
-      - id: pytest-check
-        name: pytest-check
-        entry: pytest
-        language: system
-        pass_filenames: false
-        always_run: true
