+++ pyproject.toml
@@ -15,6 +15,7 @@ pyspark = "3.1.3"
 hail = "^0.2.98"
 numpy = "^1.23.2"
 gcsfs = "^2022.8.2"
+pytest = "^7.1.3"
 
 [tool.poetry.dev-dependencies]
 pre-commit = "^2.15.0"
@@ -39,3 +40,6 @@ pytest = "^7.1.3"
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.isort]
+profile = "black"
