+++ tests/conftest.py
@@ -1,0 +1,7 @@
+import pytest
+from pyspark.sql import SparkSession
+
+
+@pytest.fixture(scope="session")
+def spark() -> SparkSession:
+    return SparkSession.builder.master("local").appName("test").getOrCreate()
